requirements.txt
aiohttp==3.8.5
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
flask==2.3.3
